fix: Fixing delay and duration functionality to be more accurate.
In Qt when using a QTimer, the default TimerType is Qt::CoarseTimer which means basically
that the QTimer will keep an accuracy within 5% of the desired interval which isn't good
enough for the Client. This fix uses the newly added class found in commit #3d3b909054db0a98314c20a442d7d4e54592ec15
to increase accuracy using delay and duration functionality and minimize code duplication in rundown widgets.

Closes: #185