fix: Configure OCS when remote triggering is enabled for GroupCommand
Previously when the "Allow Remote Triggering" checkbox was checked it only enabled the UID-label and
the UID-textbox. When the UID-textbox received a change a call to a method configuring OSC-subscriptions
was called.

Now with this fix it is possible to just enable/disable the checkbox "Allow Remote Triggering" instead of having
having to change the UID-textbox to reconfigrue OSC-subscriptions.

NOTE: This fix applies only for the GROUPCOMMAND for now. Rest of the commands will also be changed.