Fix malformed color command #293
In dd9895c118fa25c8a72a28ed8b5e7254f35d8a76 the `printf` syntax using
`%02d` was replaced with `QString` and its `arg` function. The default
fill character of `arg` is a whitespace. This broke the formatting of
the color code if they had a leading zero in any component.

In 22c838ba5fd26c1b796c8915797fc633a21415d3 this was fixed for
`Timecode.cpp`. The same fix was now also applied in the other places.